update(): restructure + configuration for common variables
