+++ src/Config.ts
@@ -1,0 +1,5 @@
+var Config = {
+  windowTitle: "WIngo - Vertretungsplan Management"
+};
+
+export { Config };
